+++ cpp/oneapi/dal/backend/primitives/optimizers/line_search_dpc.cpp
@@ -40,7 +40,7 @@ Float backtracking(sycl::queue queue,
     Float df0 = 0;
     dot_product(queue, grad_f0, direction, result.get_mutable_data(), &df0, deps + precompute)
         .wait_and_throw();
-    std::int32_t iter_num = 0;
+    std::int64_t iter_num = 0;
     Float cur_val = 0;
     while ((iter_num == 0 || cur_val > f0 + c1 * alpha * df0) && iter_num < 100) {
         if (iter_num > 0) {
